% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caption-functions.r
\name{add_index}
\alias{add_index}
\title{Add type}
\usage{
add_index(new_type)
}
\arguments{
\item{new_type}{name of new type}
}
\description{
Add a type different from "Figure" and "Table" to
  \code{typesinfo}
}

