% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeratermd-package.r
\docType{package}
\name{enumeratermd}
\alias{enumeratermd}
\alias{enumeratermd-package}
\title{Enumerate Tables and Figures in RMarkdown (rmd files)}
\description{
\code{rmd_cap} produces a caption, records a label (chunk label as default)
  and assigns a number. These number can be referenced inline with
  \code{rmd_ref} to retrieve the number.
}

