% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caption-functions.r
\name{list_of}
\alias{list_of}
\title{List of <types>}
\usage{
list_of(types = NULL)
}
\arguments{
\item{types}{what types do you want list for? (if \code{NULL} all)}
}
\description{
Create List of Tables, Figures, etc.
}

