% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeratermd-package.r
\docType{package}
\name{proh}
\alias{proh}
\alias{proh-package}
\title{Something about the package}
\description{
Details....
}
\note{
This package is an extension of the \pkg{kfigr} package and was
  essentially put together by copying and adding to the source code from
  github.
}
\references{
\url{https://github.com/mkoohafkan/kfigr}
}

