% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caption-functions.r
\name{rmd_cap}
\alias{fig_cap}
\alias{rmd_cap}
\alias{tab_cap}
\title{Capions in rmarkdown}
\usage{
rmd_cap(caption, label, type = NULL)

tab_cap(caption = "", label = NULL)

fig_cap(caption = "", label = NULL)
}
\arguments{
\item{caption}{the caption}

\item{label}{the label you can refer to in the .rmd document. If \code{NULL}
it will inherit the 'label' from the current chunk}

\item{type}{usually "Table" or "Figure". User do not have to care as you use
\code{tab_cap} or \code{fig_cap}, which calls \code{rmd_cap} with this
argument fixed}
}
\description{
This function allows enumerated tables and figures in rmarkdown.
  In connection to a table or figure you use \code{rmd_cap} (or convenience
  functions \code{tab_cap} or \code{fig_cap}) to record a label and a number
  (within the \code{typesinfo} environment). Within the text you can then
  refer to them as \code{`r rmd_ref(<label>)`}.
}
\details{
The \code{rmd_cap}, and convinience functions \code{tab_cap} and
  \code{fig_cap}, keeps track of Figures and Tables. Use e.g. \code{fig.cap =
  fig_cap("My Caption", "myLabel")} in a chunk header and \code{`r
  rmd_ref("myLab")`} in your rmd document.
}
\section{Functions}{
\itemize{
\item \code{tab_cap}: Captions for tables

\item \code{fig_cap}: Captions for figures
}}
\note{
You might have to knit your document twice for it to work. Relative
  changes or insertions of tables/figures probably requires you to reset the
  information on table/figure enumeration so far (use
  \code{index_type_reset}).
}
\seealso{
\code{\link{rmd_ref}}
}

