% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caption-functions.r
\name{rmd_ref}
\alias{rmd_ref}
\title{Get reference number}
\usage{
rmd_ref(label, type = TRUE, unknown = "*<_unset reference_>*")
}
\arguments{
\item{label}{a label}

\item{type}{look in "table" or "figure" (if \code{NULL} look everywhere)}

\item{unknown}{text if nothing found}
}
\description{
Get the refenrence number associated with a label
}

